Address remaining review findings: pins, app version, ticket dedup
- CI: new "connectcore sibling require pins" step asserts each openrung
  sibling require in connectcore/go.mod equals that sibling's VERSION
  file, and the standalone-step comment no longer claims to be the
  fetched-consumer build (the module's own replaces apply there). The
  pin/upgrade procedure documents the enforced equality and the
  partial-tag-failure recovery (re-run the idempotent tag job).
- connectcore/client: SetAppVersion lets fetched consumers (mobile
  bindings foremost) report a real app version in code, since they
  cannot inject the -X link-time symbol and would ship "dev"; the
  ldflags path is unchanged for in-repo builds. Documented in the
  README, with a test.
- connectcore/client: RequestWSSSessionTicket injects Identity and
  passes the request through whole instead of rebuilding it
  field-by-field, so a future WSSTicketRequest field cannot be silently
  zeroed; the WSSTicketStatusError alias comment now names its real
  consumer (the engine's WSS ladder, not desktop).
- internal/relay: WSSSessionTicketRequest/Response become aliases of the
  brokerapi ticket types (matching the ErrorResponse/GeoLocation
  pattern), giving the broker side the bearer-redacting String for free
  and ending the last hand-duplicated wire pair; the dead
  EffectiveNodeClass wrapper is deleted and its test moves to brokerapi,
  where the implementation lives.
- docs/architecture.md: the PunchEstablisher paragraph names
  internal/enginepunch, consistent with the module docs.
